@@ -1,23 +1,22 @@
 [project]
 name = "mcp-server-tavily"
 version = "0.1.0"
-description = "MCP Server providing a tool for Tavily web search."
+description = "MCP Server for web search using Tavily API"
 authors = [
-    {name = "Your Name", email = "your.email@example.com"}
+    { name = "XyAgent", email = "agent@xyber.inc" },
 ]
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 
 dependencies = [
-    "mcp>=1.6.0,<2.0.0",
-    "starlette>=0.37.0",
-    "uvicorn[standard]>=0.29.0",
-    "python-dotenv>=1.0.0",
-    "pydantic-settings>=2.0",
-    "langchain-tavily>=0.1.5",
-    "langchain-mcp-adapters>=0.0.9",
-    "langgraph>=0.0.20",
-    "langchain-together>=0.0.5",
+    "mcp>=1.8.0,<2.0.0",
+    "pydantic-settings>=2.0",  
+    "uvicorn[standard]",
+    "fastmcp>=2.3.4",
+    "fastapi>=0.115.12",
+    # Tavily-specific dependencies
+    "tavily-python>=0.2.0",
+    "langchain-tavily>=0.1.0",
 ]
 
 
