@@ -1,52 +1,40 @@
 import argparse
 import logging
 import os
-
 import uvicorn
-from mcp.server import Server
-from mcp.server.sse import SseServerTransport
-from mcp_server_calculator.logging_config import (configure_logging,
-                                                  logging_level)
-from mcp_server_calculator.server import server
-from starlette.applications import Starlette
-from starlette.requests import Request
-from starlette.routing import Mount, Route
+from fastapi import FastAPI
+
+from mcp_server_tavily.logging_config import (configure_logging,
+                                            logging_level)
+
+from mcp_server_tavily.server import mcp_server
 
 configure_logging()
 logger = logging.getLogger(__name__)
 
 # --- Application Factory --- #
 
+def create_app() -> FastAPI:
+    """Create a FastAPI application that can serve the provided mcp server with SSE."""
+    # Create the MCP ASGI app
+    mcp_app = mcp_server.http_app(path="/mcp", transport="streamable-http")
+    
+    # Create FastAPI app
+    app = FastAPI(
+        title="Tavily MCP Server",
+        description="MCP server for web search using Tavily API",
+        version="1.0.0",
+        lifespan=mcp_app.router.lifespan_context
+    )   
+    
+    # Mount MCP server
+    app.mount("/mcp-server", mcp_app)
 
-def create_starlette_app() -> Starlette:
-    """Create a Starlette application that can server the provied mcp server with SSE."""
+    return app
 
-    sse = SseServerTransport("/messages/")
-    mcp_server: Server = server
-
-    async def handle_sse(request: Request) -> None:
-        async with sse.connect_sse(
-            request.scope,
-            request.receive,
-            request._send,  # noqa: SLF001
-        ) as (read_stream, write_stream):
-            await mcp_server.run(
-                read_stream,
-                write_stream,
-                mcp_server.create_initialization_options(),
-            )
-
-    return Starlette(
-        debug=logging_level == "DEBUG",
-        routes=[
-            Route("/sse", endpoint=handle_sse),
-            Mount("/messages/", app=sse.handle_post_message),
-        ],
-    )
 
 if __name__ == "__main__":
-    # --- Command Line Argument Parsing ---
-    parser = argparse.ArgumentParser(description='Run MCP Starlette server for Tavily Web Search')
+    parser = argparse.ArgumentParser(description="Run Tavily MCP server")
     parser.add_argument(
         "--host",
         default=os.getenv("MCP_TAVILY_HOST", "0.0.0.0"),
@@ -67,11 +55,10 @@ if __name__ == "__main__":
     )
 
     args = parser.parse_args()
-    logger.info(f"Starting MCP Tavily Server on http://{args.host}:{args.port}")
-    
-    # Run Uvicorn server
+    logger.info(f"Starting Tavily MCP server on {args.host}:{args.port}")
+
     uvicorn.run(
-        "mcp_server_tavily.__main__:create_starlette_app",
+        "mcp_server_tavily.__main__:create_app",
         host=args.host,
         port=args.port,
         reload=args.reload,
