@@ -1,7 +1,7 @@
 """Tavily client module for the MCP server."""
 
 from mcp_server_tavily.tavily.module import _TavilyService, get_tavily_service, TavilySearchResult
-from mcp_server_tavily.tavily.config import TavilyConfig, TavilyServiceError
+from mcp_server_tavily.tavily.config import TavilyConfig, TavilyServiceError, TavilyApiError, TavilyConfigError
 
 __all__ = [
     "_TavilyService",
@@ -9,4 +9,6 @@ __all__ = [
     "TavilySearchResult",
     "TavilyConfig",
     "TavilyServiceError",
+    "TavilyApiError",
+    "TavilyConfigError",
 ] 
