@@ -4,77 +4,63 @@
 import argparse
 import logging
 import os
-
 import uvicorn
-from mcp.server import Server
-from mcp.server.sse import SseServerTransport
+from fastapi import FastAPI
+
 from mcp_server_twitter.logging_config import (configure_logging,
                                                   logging_level)
-from mcp_server_twitter.server import server
-from starlette.applications import Starlette
-from starlette.requests import Request
-from starlette.routing import Mount, Route
+from mcp_server_twitter.server import mcp_server
 
 configure_logging()
 logger = logging.getLogger(__name__)
 
 # --- Application Factory --- #
 
+def create_app() -> FastAPI:
+    """Create a FastAPI application that can serve the provided mcp server with SSE."""
+    # Create the MCP ASGI app
+    mcp_app = mcp_server.http_app(path="/mcp", transport="streamable-http")
+    
+    # Create FastAPI app with proper FastMCP lifespan
+    app = FastAPI(
+        title="Twitter MCP Server",
+        description="MCP server for Twitter integration",
+        version="1.0.0",
+        lifespan=mcp_app.lifespan
+    )   
+    
+    # Mount MCP server
+    app.mount("/mcp-server", mcp_app)
 
-def create_starlette_app() -> Starlette:
-    """Create a Starlette application that can server the provied mcp server with SSE."""
-
-    sse = SseServerTransport("/messages/")
-    mcp_server: Server = server
-
-    async def handle_sse(request: Request) -> None:
-        async with sse.connect_sse(
-            request.scope,
-            request.receive,
-            request._send,  # noqa: SLF001
-        ) as (read_stream, write_stream):
-            await mcp_server.run(
-                read_stream,
-                write_stream,
-                mcp_server.create_initialization_options(),
-            )
-
-    return Starlette(
-        debug=logging_level == "DEBUG",
-        routes=[
-            Route("/sse", endpoint=handle_sse),
-            Mount("/messages/", app=sse.handle_post_message),
-        ],
-    )
+    return app
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="Run Twitter MCP server")
     parser.add_argument(
         "--host",
-        default=os.getenv("MCP_TWITTER_HOST", "0.0.0.0"),  # Override with your env variables
-        help="Host to bind to (Default: MCP_TWITTER_HOST or 0.0.0.0)", # Override with your env variables
+        default=os.getenv("MCP_TWITTER_HOST", "0.0.0.0"),
+        help="Host to bind to (Default: MCP_TWITTER_HOST or 0.0.0.0)",
     )
     parser.add_argument(
         "--port",
         type=int,
-        default=int(os.getenv("MCP_TWITTER_PORT", "8000")), # Override with your env variables
-        help="Port to listen on (Default: MCP_TWITTER_PORT or 8000)", # Override with your env variables
+        default=int(os.getenv("MCP_TWITTER_PORT", "8000")),
+        help="Port to listen on (Default: MCP_TWITTER_PORT or 8000)",
     )
     parser.add_argument(
         "--reload",
         action="store_true",
-        default=os.getenv("MCP_CALCULATOR_HOT_RELOAD", "false").lower()
-        in ("true", "1", "t", "yes"), # Override with your env variables
-        help="Enable hot reload (env: MCP_CALCULATOR_HOT_RELOAD)", # Override with your env variables
+        default=os.getenv("TWITTER_HOT_RELOAD", "false").lower()
+        in ("true", "1", "t", "yes"),
+        help="Enable hot reload (env: TWITTER_HOT_RELOAD)",
     )
 
     args = parser.parse_args()
-    logger.info(f"Starting TWITTER MCP server on {args.host}:{args.port}")
+    logger.info(f"Starting Twitter MCP server on {args.host}:{args.port}")
 
-    # Don't forget to change the module name to your own!
     uvicorn.run(
-        "mcp_server_twitter.__main__:create_starlette_app",
+        "mcp_server_twitter.__main__:create_app",
         host=args.host,
         port=args.port,
         reload=args.reload,
