@@ -1,7 +1,7 @@
-from mcp_server_twitter.twitter.module import TwitterClient, get_twitter_client
+from mcp_server_twitter.twitter.module import AsyncTwitterClient, get_twitter_client
 
 __all__ = [
-    "TwitterClient",
+    "AsyncTwitterClient",
     "get_twitter_client",
     "TwitterConfig",
 ]
