@@ -1,40 +1,17 @@
-from __future__ import annotations
-
 import logging
 from contextlib import asynccontextmanager
-from enum import StrEnum
+
 from typing import AsyncIterator, Any, Literal
 from fastmcp import Context, FastMCP
 from fastmcp.exceptions import ToolError
 
-from mcp.server import Server
-from mcp.types import TextContent, Tool
 from mcp_server_youtube.youtube import (YouTubeClientError, YouTubeSearcher,
                                         get_youtube_searcher)
-from pydantic import BaseModel, Field, ValidationError
 
 logger = logging.getLogger(__name__)
 
 
-# --- Tool Constants & Enums ---
-class YouTubeToolNames(StrEnum):
-    """Enum for YouTube MCP tool names."""
 
-    SEARCH_AND_TRANSCRIPT = "youtube_search_and_transcript"
-
-
-# --- Tool Input/Output Schemas --- #
-class YouTubeSearchRequest(BaseModel):
-    """Input schema for the youtube_search_and_transcript tool."""
-
-    query: str = Field(..., description="The search query string for YouTube videos.")
-    max_results: int = Field(
-        default=3, description="Maximum number of video results to return.", ge=1, le=20
-    )
-    transcript_language: str = Field(
-        default="en",
-        description="The language code for the transcript (e.g., 'en', 'es').",
-    )
 
 
 # --- Lifespan Management --- #
